+++ spigot/spigot18/src/com/elikill58/negativity/spigot18/Spigot_1_18_R1.java
@@ -159,7 +159,7 @@ public class Spigot_1_18_R1 extends SpigotVersionAdapter {
 			return new NPacketHandshakeInSetProtocol(packet.getProtocolVersion(), packet.hostName, packet.port);
 		});
 		
-		negativityToPlatform.put(PacketType.Server.PING, (p, f) -> new ClientboundPingPacket(((NPacketPlayOutPing) f).id));
+		negativityToPlatform.put(PacketType.Server.PING, (p, f) -> new ClientboundPingPacket((int) ((NPacketPlayOutPing) f).id));
 		
 		log();
 	}
